Handled the rejection or the cancellation of a profile to a conference

+++ app/Http/Controllers/ConferenceVehicleController.php
@@ -35,16 +35,28 @@ class ConferenceVehicleController extends Controller
     public function store($conferences, Request $request)
     {
         try {
+            // Check if the Conference exists.
+            $conf = Conference::find($conferences);
+            if (!$conf) {
+                return response()->error(404);
+            }
+
+            //Check if conference manager belongs to this conference OR admin
+            $userId = $request->header('ID');
+            if (!$conf->managers()->where('user_id', $userID)->get() ||
+                Sentinel::findById($userId)->roles()->first()->name != 'System Administrator') {
+                return response()->error("403" , "Permission Denied");
+            }
+
             $vehicle = Vehicle::create($request->all());
 
             //Storing new objects into pivot table.
             //Creating the pivot between conf & vehicle
-            Conference::find($conferences)
-                        ->vehicles()
-                        ->attach($vehicle, ['type' => $request->type]);
+            $conf->vehicles()
+                 ->attach($vehicle, ['type' => $request->type]);
             return response()->success();
         } catch (Exception $e) {
-            return response()->error("Viper" , $e);
+            return response()->error(500 , $e);
         }
     }
 
@@ -64,7 +76,7 @@ class ConferenceVehicleController extends Controller
             }
             return $vehicle;
         } catch (Exception $e) {
-            return response()->error("Vine" , $e);
+            return response()->error(500 , $e);
         }
     }
 
@@ -78,6 +90,19 @@ class ConferenceVehicleController extends Controller
     public function update($conferences, Request $request)
     {
         try {
+            // Check if the Conference exists.
+            $conf = Conference::find($conferences);
+            if (!$conf) {
+                return response()->error(404);
+            }
+
+            //Check if conference manager belongs to this conference OR admin
+            $userId = $request->header('ID');
+            if (!$conf->managers()->where('user_id', $userID)->get() ||
+                Sentinel::findById($userId)->roles()->first()->name != 'System Administrator') {
+                return response()->error("403" , "Permission Denied");
+            }
+
             $newVehicleData = array (
                 'name' => $request->name,
                 'passenger_count' => $request->passenger_count,
@@ -86,9 +111,8 @@ class ConferenceVehicleController extends Controller
             Vehicle::where('id', $id)->update($newVehicleData);
 
             //Updating Pivot Table
-            Conference::find($conferences)
-                        ->vehicles()
-                        ->updateExistingPivot($id,['type' => $request->type]);
+            $conf->vehicles()
+                 ->updateExistingPivot($id,['type' => $request->type]);
             return response()->success();
         } catch (Exception $e) {
             return response()->error("500" , $e);
@@ -104,14 +128,31 @@ class ConferenceVehicleController extends Controller
     public function destroy($conferences, $vehicles)
     {
         try {
-            $vehicle = Vehicle::findorfail($vehicles);
+            // Check if the Conference exists.
+            $conf = Conference::find($conferences);
+            if (!$conf) {
+                return response()->error(404);
+            }
+
+            //Check if conference manager belongs to this conference OR admin
+            $userId = $request->header('ID');
+            if (!$conf->managers()->where('user_id', $userID)->get() ||
+                Sentinel::findById($userId)->roles()->first()->name != 'System Administrator') {
+                return response()->error("403" , "Permission Denied");
+            }
+
+
+            $vehicle = Vehicle::find($vehicles);
+            if (!$vehicle) {
+                return response()->error(404);
+            }
+
             if ($vehicle->passengers()->count()){
                 return response()->error("409" , "Passengers still in this Vehicle");
             }
 
             //Remove the Pivot Row - From Conference/Event - Vehicles
-            Vehicle::find($vehicles)
-                    ->conferences()
+            $vehicle->conferences()
                     ->detach();
 
             Vehicle::destroy($vehicles);
